refactor: Move render pagination util to UI library
